refactor: rename files where appropriate
