+++ personality-insights/v3.ts
