+++ test/unit/discovery.v1.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const helper = require('ibm-cloud-sdk-core');
-const DiscoveryV1 = require('../../discovery/v1-generated');
+const DiscoveryV1 = require('../../discovery/v1');
 const utils = require('../resources/unitTestUtils');
 
 const getOptions = utils.getOptions;
