+++ visual-recognition/v3.ts
